[#930] reorganised test packages
