+++ app/src/androidTest/java/org/akvo/flow/activity/form/formfill/OptionsQuestionViewMultipleTest.java
@@ -18,24 +18,23 @@
  *
  */
 
-package org.akvo.flow.activity.form.questionviews;
+package org.akvo.flow.activity.form.formfill;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
 import android.support.test.InstrumentationRegistry;
-import android.support.test.espresso.Espresso;
+import android.support.test.espresso.ViewInteraction;
 import android.support.test.filters.MediumTest;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
+import android.view.View;
+import android.widget.CheckBox;
 
 import org.akvo.flow.R;
 import org.akvo.flow.activity.FormActivity;
 import org.akvo.flow.activity.form.data.SurveyInstaller;
 import org.akvo.flow.activity.form.data.SurveyRequisite;
-import org.akvo.flow.domain.Survey;
-import org.akvo.flow.domain.ValidationRule;
+import org.hamcrest.core.IsInstanceOf;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -43,35 +42,36 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.io.IOException;
-
 import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.typeText;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.hasErrorText;
+import static android.support.test.espresso.matcher.ViewMatchers.isChecked;
+import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static org.akvo.flow.activity.Constants.TEST_FORM_SURVEY_INSTANCE_ID;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.clickNext;
-import static org.akvo.flow.activity.form.FormActivityTestUtil.fillFreeTextQuestion;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.getFormActivityIntent;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.verifyQuestionTitleDisplayed;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonDisabled;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonEnabled;
-import static org.akvo.flow.tests.R.raw.number_form;
+import static org.akvo.flow.tests.R.raw.option_multiple_other_form;
+import static org.hamcrest.Matchers.allOf;
 
 @MediumTest
 @RunWith(AndroidJUnit4.class)
-public class NumberQuestionViewTest {
+public class OptionsQuestionViewMultipleTest {
 
+    private static final String FORM_TITLE = "OptionsQuestionForm";
     private static SurveyInstaller installer;
-    private static Survey survey;
 
     @Rule
     public ActivityTestRule<FormActivity> rule = new ActivityTestRule<FormActivity>(
             FormActivity.class) {
         @Override
         protected Intent getActivityIntent() {
-            return getFormActivityIntent(47323002L, "49783002", "NumberForm", 0L, false);
+            return getFormActivityIntent(42573002L, "43623002", FORM_TITLE, 0L, false);
         }
     };
 
@@ -80,7 +80,7 @@ public class NumberQuestionViewTest {
         Context targetContext = InstrumentationRegistry.getTargetContext();
         SurveyRequisite.setRequisites(targetContext);
         installer = new SurveyInstaller(targetContext);
-        survey = installer.installSurvey(number_form, InstrumentationRegistry.getContext());
+        installer.installSurvey(option_multiple_other_form, InstrumentationRegistry.getContext());
     }
 
     @After
@@ -95,76 +95,81 @@ public class NumberQuestionViewTest {
     }
 
     @Test
-    public void ensureCanFillNumberQuestion() throws IOException {
-        fillNumberQuestion(50);
+    public void ensureCanFillOneOptionsQuestion() throws Exception {
+        verifyQuestionTitleDisplayed();
+
+        fillOptionsQuestion(0);
+
+        verifyOptionSelected(0);
+
         clickNext();
+
         verifySubmitButtonEnabled();
     }
 
+    @Test
+    public void ensureCanFillMultipleOptionsQuestion() throws Exception {
+        verifyQuestionTitleDisplayed();
+
+        fillOptionsQuestion(0);
+        fillOptionsQuestion(1);
+
+        verifyOptionSelected(0);
+        verifyOptionSelected(1);
+
+        clickNext();
+
+        verifySubmitButtonEnabled();
+    }
 
     @Test
-    public void ensureCanNotifyWrongMaxInputNumberQuestion() throws Exception {
-        fillNumberQuestion(2000);
+    public void ensureCanFillOtherOptionsQuestion() throws Exception {
+        verifyQuestionTitleDisplayed();
 
-        verifyNumberTooLargeErrorShown();
+        fillOptionsQuestion(2);
+        fillOtherValue("other option");
+
+        verifyOtherOption(2, "other option");
+
         clickNext();
+
+        verifySubmitButtonEnabled();
+    }
+
+    private void fillOtherValue(String text) {
+        onView(withId(R.id.other_option_input)).perform(typeText(text));
+        onView(withId(android.R.id.button1)).perform(click());
+    }
+
+    private void verifyOtherOption(int option, String text) {
+        ViewInteraction otherOption = getCheckbox(option);
+        otherOption.check(matches(isChecked()));
+        onView(withId(R.id.other_option_text))
+                .check(matches(allOf(isDisplayed(), withText(text))));
+    }
+
+    @Test
+    public void ensureCannotSubmitIfNoOptionSelected() throws Exception {
+        verifyQuestionTitleDisplayed();
+
+        clickNext();
+
         verifySubmitButtonDisabled();
     }
 
-    @Test
-    public void ensureCanNotifyWrongMinInputNumberQuestion() throws Exception {
-        fillNumberQuestion(0);
-
-        verifyNumberTooSmallErrorShown();
-        clickNext();
-        verifySubmitButtonDisabled();
+    private void verifyOptionSelected(int option) {
+        ViewInteraction singleChoiceOption = getCheckbox(option);
+        singleChoiceOption.check(matches(isChecked()));
     }
 
-    @Test
-    public void ensureCannotEnterText() throws Exception {
-        fillFreeTextQuestion("This is an answer to your question");
-
-        onView(withId(R.id.input_et)).check(matches(withText("")));
+    private void fillOptionsQuestion(int option) {
+        ViewInteraction checkbox = getCheckbox(option);
+        checkbox.perform(click());
     }
 
-    @Test
-    public void ensureCannotEnterSigned() throws Exception {
-        fillFreeTextQuestion("-1");
-
-        onView(withId(R.id.input_et)).check(matches(withText("1")));
+    private ViewInteraction getCheckbox(int option) {
+        return onView(allOf(withId(option), IsInstanceOf.<View>instanceOf(CheckBox.class)));
     }
 
-    @Test
-    public void ensureCannotEnterDecimal() throws Exception {
-        fillFreeTextQuestion("1.1");
 
-        onView(withId(R.id.input_et)).check(matches(withText("11")));
-    }
-
-    @NonNull
-    private String getString(@StringRes int stringResId) {
-        return rule.getActivity().getApplicationContext().getResources()
-                .getString(stringResId);
-    }
-
-    private void verifyNumberTooLargeErrorShown() {
-        int maxValue = getValidationRule().getMaxVal().intValue();
-        String tooLargeError = getString(R.string.toolargeerr);
-        onView(withId(R.id.input_et)).check(matches(hasErrorText(tooLargeError + maxValue)));
-    }
-
-    private void verifyNumberTooSmallErrorShown() {
-        int minValue = getValidationRule().getMinVal().intValue();
-        String tooSmallError = getString(R.string.toosmallerr);
-        onView(withId(R.id.input_et)).check(matches(hasErrorText(tooSmallError + minValue)));
-    }
-
-    private ValidationRule getValidationRule() {
-        return survey.getQuestionGroups().get(0).getQuestions().get(0).getValidationRule();
-    }
-
-    private void fillNumberQuestion(int firstValue) throws IOException {
-        onView(withId(R.id.input_et)).perform(typeText(String.valueOf(firstValue)));
-        Espresso.closeSoftKeyboard();
-    }
 }
