+++ app/src/androidTest/java/org/akvo/flow/activity/form/formfill/OptionsQuestionViewSingleTest.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.akvo.flow.activity.form.questionviews;
+package org.akvo.flow.activity.form.formfill;
 
 import android.content.Context;
 import android.content.Intent;
@@ -28,9 +28,8 @@ import android.support.test.filters.MediumTest;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.view.View;
-import android.widget.CheckBox;
+import android.widget.RadioGroup;
 
-import org.akvo.flow.R;
 import org.akvo.flow.activity.FormActivity;
 import org.akvo.flow.activity.form.data.SurveyInstaller;
 import org.akvo.flow.activity.form.data.SurveyRequisite;
@@ -44,24 +43,23 @@ import org.junit.runner.RunWith;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.typeText;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isChecked;
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static android.support.test.espresso.matcher.ViewMatchers.isDescendantOfA;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static org.akvo.flow.activity.Constants.TEST_FORM_SURVEY_INSTANCE_ID;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.clickNext;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.getFormActivityIntent;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.verifyQuestionTitleDisplayed;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonDisabled;
 import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonEnabled;
-import static org.akvo.flow.tests.R.raw.option_multiple_other_form;
+import static org.akvo.flow.tests.R.raw.option_form;
 import static org.hamcrest.Matchers.allOf;
+import static org.hamcrest.core.IsNot.not;
 
 @MediumTest
 @RunWith(AndroidJUnit4.class)
-public class OptionsQuestionViewMultipleTest {
+public class OptionsQuestionViewSingleTest {
 
     private static final String FORM_TITLE = "OptionsQuestionForm";
     private static SurveyInstaller installer;
@@ -80,7 +78,7 @@ public class OptionsQuestionViewMultipleTest {
         Context targetContext = InstrumentationRegistry.getTargetContext();
         SurveyRequisite.setRequisites(targetContext);
         installer = new SurveyInstaller(targetContext);
-        installer.installSurvey(option_multiple_other_form, InstrumentationRegistry.getContext());
+        installer.installSurvey(option_form, InstrumentationRegistry.getContext());
     }
 
     @After
@@ -95,57 +93,17 @@ public class OptionsQuestionViewMultipleTest {
     }
 
     @Test
-    public void ensureCanFillOneOptionsQuestion() throws Exception {
+    public void ensureCanFillOptionsQuestion() throws Exception {
         verifyQuestionTitleDisplayed();
 
         fillOptionsQuestion(0);
 
-        verifyOptionSelected(0);
+        verifyOptionSelected();
+        verifyOtherOptionUnselected();
 
         clickNext();
 
         verifySubmitButtonEnabled();
-    }
-
-    @Test
-    public void ensureCanFillMultipleOptionsQuestion() throws Exception {
-        verifyQuestionTitleDisplayed();
-
-        fillOptionsQuestion(0);
-        fillOptionsQuestion(1);
-
-        verifyOptionSelected(0);
-        verifyOptionSelected(1);
-
-        clickNext();
-
-        verifySubmitButtonEnabled();
-    }
-
-    @Test
-    public void ensureCanFillOtherOptionsQuestion() throws Exception {
-        verifyQuestionTitleDisplayed();
-
-        fillOptionsQuestion(2);
-        fillOtherValue("other option");
-
-        verifyOtherOption(2, "other option");
-
-        clickNext();
-
-        verifySubmitButtonEnabled();
-    }
-
-    private void fillOtherValue(String text) {
-        onView(withId(R.id.other_option_input)).perform(typeText(text));
-        onView(withId(android.R.id.button1)).perform(click());
-    }
-
-    private void verifyOtherOption(int option, String text) {
-        ViewInteraction otherOption = getCheckbox(option);
-        otherOption.check(matches(isChecked()));
-        onView(withId(R.id.other_option_text))
-                .check(matches(allOf(isDisplayed(), withText(text))));
     }
 
     @Test
@@ -157,18 +115,24 @@ public class OptionsQuestionViewMultipleTest {
         verifySubmitButtonDisabled();
     }
 
-    private void verifyOptionSelected(int option) {
-        ViewInteraction singleChoiceOption = getCheckbox(option);
+    private void verifyOptionSelected() {
+        ViewInteraction singleChoiceOption = getSingleChoiceRadioButton(0);
         singleChoiceOption.check(matches(isChecked()));
     }
 
-    private void fillOptionsQuestion(int option) {
-        ViewInteraction checkbox = getCheckbox(option);
-        checkbox.perform(click());
+    private void verifyOtherOptionUnselected() {
+        ViewInteraction singleChoiceOption = getSingleChoiceRadioButton(1);
+        singleChoiceOption.check(matches(not(isChecked())));
     }
 
-    private ViewInteraction getCheckbox(int option) {
-        return onView(allOf(withId(option), IsInstanceOf.<View>instanceOf(CheckBox.class)));
+    private void fillOptionsQuestion(int option) {
+        ViewInteraction radioButton = getSingleChoiceRadioButton(option);
+        radioButton.perform(click());
+    }
+
+    private ViewInteraction getSingleChoiceRadioButton(int option) {
+        return onView(allOf(withId(option),
+                isDescendantOfA(IsInstanceOf.<View>instanceOf(RadioGroup.class))));
     }
